Add explicit type conversions where necessary: device/
These are cases that are implicitly narrowing today, and must do so
explicitly in order to enable -Wc++11-narrowing.  No behavior change
intended.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2947537
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891362}

+++ device/bluetooth/test/fake_bluetooth_le_advertisement_winrt.cc
@@ -206,7 +206,7 @@ HRESULT FakeBluetoothLEAdvertisementWinrt::GetSectionsByType(
   std::vector<ComPtr<IBluetoothLEAdvertisementDataSection>> data_sections;
   if (type == BluetoothDeviceWinrt::kTxPowerLevelDataSection && tx_power_) {
     data_sections.push_back(Make<FakeBluetoothLEAdvertisementDataSectionWinrt>(
-        std::vector<uint8_t>({*tx_power_})));
+        std::vector<uint8_t>({static_cast<uint8_t>(*tx_power_)})));
   }
 
   // For simplicity we only implement querying 128 Bit UUID Service Data.
